change the get for all rentals to return the all the properties of a CarDto

+++ src/main/java/academy/mindswap/rentacar/service/CarServiceImpl.java
@@ -4,6 +4,7 @@ import academy.mindswap.rentacar.converter.CarConverter;
 import academy.mindswap.rentacar.dto.CarCreateDto;
 import academy.mindswap.rentacar.dto.CarDto;
 import academy.mindswap.rentacar.dto.UserDto;
+import academy.mindswap.rentacar.exceptions.CarDoesntExists;
 import academy.mindswap.rentacar.model.Car;
 import academy.mindswap.rentacar.model.User;
 import academy.mindswap.rentacar.repository.CarRepository;
@@ -42,8 +43,11 @@ public class CarServiceImpl implements CarService{
     }
 
     @Override
-    public CarDto getCarById(Long userId) {
-        Car car = carRepository.getReferenceById(userId);
+    public CarDto getCarById(Long carId) {
+        if (!carRepository.existsById(carId)){
+            throw new CarDoesntExists("Car doesn't Exists");
+        }
+        Car car = carRepository.getReferenceById(carId);
         return carConverter.fromCarEntityToCarDto(car);
     }
 
@@ -58,6 +62,10 @@ public class CarServiceImpl implements CarService{
 
     @Override
     public CarDto updateCar(Long id, CarDto carDto) {
+        if (!carRepository.existsById(id)){
+            throw new CarDoesntExists("Car doesn't Exists");
+        }
+
         Car car = carRepository.getReferenceById(id);
         car.setBrand(carDto.getBrand());
         car.setModel(carDto.getModel());
@@ -67,6 +75,10 @@ public class CarServiceImpl implements CarService{
 
     @Override
     public void deleteCar(Long carId) {
+        if (!carRepository.existsById(carId)){
+            throw new CarDoesntExists("Car doesn't Exists");
+        }
+
         carRepository.deleteById(carId);
 
     }
